violet/pyproject.toml
[project]
name = "violet"
version = "0.1.0"
description = "next generation interaction system"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache License 2.0"}
authors = [
    {name = "ClearXs", email = "jiangw1027@gmail.com"}
]
maintainers = [
    {name = "ClearXs", email = "jiangw1027@gmail.com"}
]
dependencies = [
    "anthropic>=0.63.0",
    "cn2an>=0.5.23",
    "colorama>=0.4.6",
    "demjson3>=3.0.6",
    "docstring-parser>=0.17.0",
    "fast-langdetect>=0.3.1",
    "fastapi[standard]>=0.115.2",
    "ffmpeg-python>=0.2.0",
    "funasr==1.0.27",
    "g2p-en>=2.1.0",
    "gradio<5",
    "httpx-sse>=0.4.1",
    "humps>=0.2.2",
    "jieba>=0.42.1",
    "jieba-fast>=0.53",
    "librosa==0.10.2",
    "llama-cpp-python>=0.3.15",
    "llama-index>=0.13.1",
    "matplotlib>=3.10.5",
    "numpy<2.0",
    "pathvalidate>=3.3.1",
    "peft>=0.17.0",
    "pydantic-settings>=2.10.1",
    "pyopenjtalk>=0.4.1",
    "pypinyin>=0.55.0",
    "pytorch-lightning>=2.4",
    "pytz>=2025.2",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.13.0",
    "split-lang>=2.1.0",
    "torchaudio>=2.8.0",
    "torchmetrics<=1.5",
    "wordsegment>=1.3.1",
    "x-transformers>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
]
voice = [
    "SpeechRecognition",
    "pydub",
]
full = [
    "SpeechRecognition",
    "pydub",
]


[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-tornasync>=0.6.0.post2",
]
